backend/requirements.txt
pandas==2.2.3
pybaseball==2.2.7
python-dotenv==1.0.1
requests==2.32.3
supabase==2.10.0
pytest==8.3.5
freezegun==1.5.1
